Refactor(python) Cleanup setup script. Put custom Setuptools commands to separate file
