+++ python/sdist/MANIFEST.in
@@ -6,4 +6,5 @@ include amici/*
 include amici/src/*template*
 include amici/swig/CMakeLists_model.txt
 include setup_clibs.py
+include custom_commands.py
 include version.txt
